Added LeftPegLeave and RightPegLeave
In this commit, two new commands were added: LeftPegLeave, and RightPegLeave. They essentially work the same as Middle Peg Leave Left and Right, as they go across the field once having placed the gear on the peg. The movement of the active gear intake was re-enabled, and mroe autonomous tuning was done. Tagged.